move more methods to EAMixins (#21782)
